Remove util.py and create to distribute them into util/*.py
